#0006
Finally I could make Kafka to work fine.
The problem was that when both consumer and producer are in same project the properties must be set on each one instead of global.
This misleading confusing was happening that server could not correctly process and save it.

All the point is that I need another consumer (or use docker embedded) to check the queues and example the consuming.

Signed-off-by: Isaac Alencar <44655568+Hakzai@users.noreply.github.com>